feat: dynamic reviewer filtering for votesHistory.
