+++ server/src/use-cases/version-control/azure-devops/code-review/vote-metrics/vote-metrics.unit.test.js
@@ -74,62 +74,3 @@ describe('VoteMetrics~getPullRequestVotesTimeline - method to get pull request v
     ]);
   });
 });
-
-describe('VoteMetrics~getPullRequestVotesHistory - method to get pull request votes history from votes history timeline.', () => {
-  it('should return the history of pull request votes from votes history timeline.', () => {
-    const votesCycle = [
-      {
-        id: '1',
-        author: 'Author1',
-        timeOfVote: '2022-01-01T00:00:00Z',
-        vote: 'approved',
-      },
-      {
-        id: '2',
-        author: 'Author2',
-        timeOfVote: '2022-01-01T00:00:00Z',
-        vote: 'approvedWithSuggestions',
-      },
-      {
-        id: '3',
-        author: 'Author3',
-        timeOfVote: '2022-01-01T00:00:00Z',
-        vote: 'rejected',
-      },
-      {
-        id: '4',
-        author: 'Author4',
-        timeOfVote: '2022-01-01T00:00:00Z',
-        vote: 'waitForAuthor',
-      },
-      {
-        id: '3',
-        author: 'Author3',
-        timeOfVote: '2022-01-01T00:00:00Z',
-        vote: 'approved',
-      },
-    ];
-
-    const result = VoteMetrics.getPullRequestVotesHistory(votesCycle);
-
-    expect(result).toEqual({
-      approved: 2,
-      approvedWithSuggestions: 1,
-      waitForAuthor: 1,
-      rejected: 1,
-    });
-  });
-
-  it('should return all zeros if there are no votes', () => {
-    const votesCycle = [];
-
-    const result = VoteMetrics.getPullRequestVotesHistory(votesCycle);
-
-    expect(result).toEqual({
-      approved: 0,
-      approvedWithSuggestions: 0,
-      waitForAuthor: 0,
-      rejected: 0,
-    });
-  });
-});
